requirements.txt
pynput
pystray
Pillow
XInput-Python
customtkinter
